requirements.txt
pathlib
streamlit
openai
pymupdf
python-dotenv
faiss-cpu